embedded_3dgfx

Module color_gradient

Multi-stop color gradient sampling for no_std embedded systems.

Inspired by Fyrox’s ColorGradient, adapted for static flash ROM storage and Rgb565 / embedded-graphics color types.

Example

use embedded_3dgfx::color_gradient::{ColorGradient, GradientStop};
use embedded_graphics_core::pixelcolor::Rgb565;

static FIRE_GRADIENT: ColorGradient<'static, Rgb565> = ColorGradient::new(&[
    GradientStop::new(0.0, Rgb565::new(31, 63, 31)), // White hot
    GradientStop::new(0.2, Rgb565::new(31, 63, 0)),  // Yellow
    GradientStop::new(0.6, Rgb565::new(31, 20, 0)),  // Orange/Red
    GradientStop::new(1.0, Rgb565::new(4, 4, 4)),    // Dark smoke
]);

let mid_color = FIRE_GRADIENT.sample(0.5);

Structs

ColorGradient

A multi-stop color gradient backed by a slice of GradientStop.

GradientStop

A single color key in a gradient at a normalized position [0.0, 1.0].