embedded_3dgfx::config

Trait CapSink

pub trait CapSink {
    // Required methods
    fn set_caps(&mut self, caps: ProfileCaps);
    fn clear_caps(&mut self);
    fn apply_render_defaults(&mut self, defaults: RenderDefaults);
}

Sink for the render-cap setters that apply_default_caps drives.

Implemented by the engine. Keeping this a trait means the config layer does not have to depend on the engine layer.

Required Methods

fn set_caps(&mut self, caps: ProfileCaps)

fn clear_caps(&mut self)

fn apply_render_defaults(&mut self, defaults: RenderDefaults)

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".