embedded_3dgfx

Module simplex_stroke_font

Static Simplex Hershey font data used by text gizmos.

Adapted from Bevy (bevy_gizmos::simplex_stroke_font), licensed under MIT / Apache-2.0. Original Hershey vector font data is in the public domain (https://paulbourke.net/dataformats/hershey/).

Structs

StrokeFont

A stroke font containing glyphs for the 95 printable ASCII codes (32..=126).

Constants

LINE_HEIGHT

SIMPLEX_CAP_HEIGHT

SIMPLEX_DESCENDER_DEPTH

SIMPLEX_GLYPHS

SIMPLEX_HEIGHT

SIMPLEX_POSITIONS

SIMPLEX_STROKES

SIMPLEX_STROKE_FONT

TAB_WIDTH_IN_SPACES